backend/requirements.txt
fastapi
uvicorn
google-generativeai
pydub
PyMuPDF
speechrecognition
python-multipart